+++ database/migrations/2024_03_26_200542_add_mime_type_to_mist_stream_wildcards_table.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+      Schema::table('mist_stream_wildcards', function (Blueprint $table) {
+        $table->string('mime_type')->nullable()->after('source')->comment('MIME type of the stream or push');
+      });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+      Schema::table('mist_stream_wildcards', function (Blueprint $table) {
+        $table->dropColumn('mime_type');
+      });
+    }
+};